chore: remove useless event listener
Every time a debugging session is initialized, we override the previous
`exceptions_options` with the defaults from the adapter. For this hook
to take effect, a user would have to update the configuration between an
initialize request and a launch request, but that's extremely unlikely.
Otherwise, we'll just update the default configuration with the...
defaults.